backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular
channels
channels-redis
django-cors-headers
redis
selenium
pandas
scikit-learn
xgboost
joblib
requests
